# Support NR being a multiple of vector width in int8 GEMM